Correcting task 6

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,8 +3,8 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     '''function sum_mixed_list which takes a list
     mxd_lst of integers and floats and returns their
     sum as a float.'''
-    return sum(mxd_list)
+    return float(sum(mxd_list))
